Update and rename assign2.ipynb to Shivanee_230974_assign2.ipynb
Feature importance is assessed using Lasso regression due to its inherent ability to perform feature selection. By examining the non-zero coefficients in the lasso, we can determine which features have the most substantial impact on the target variable (heating load) and some more suggestions for energy efficiency is added in the code.